+++ CHANGELOG.md
@@ -60,6 +60,7 @@
   * Merge requests can now be filtered by search scopes.
   * Merge requests can now trigger merge status rechecks when listing.
   * Merge requests can now be filtered by reviewer.
+  * Merge requests can now be filtered by environment status.
 
 # v0.1308.0